Initial work on the propagators module.
Still unsure how to handle the different kinds of RKs (Verner, Cash,
original, etc.) and their stages. Should the number of stages be in the
initialization of the RK, or should there be one implementation per RK
stage? The latter is easier to code (and execute quicker), but seems
like it isn't the best design. Moreover, currently the adaptive or fixed
step is defined in the options: this means calling an RK45 with a fixed
step will actually execute an RK4 (user should probably be warned?):
that can be confusing.