Drastically changed paradigm after thinking about similarities
Needs thorough testing. Also needs the definition of the RKs as Butcher
tables. But the good news now is that `Propagator` is the main struct to
handle any RK which is consistent.

TODO: Adaptive step size; Massive amounts of documentation (which will
come after testing); Massive benchmarking to ensure that this code is
both robust and especially fast.