Ferhlberg65 seems broken, need a new source on the coefficients.
I tried using an embedded version of RK Ferhlberg65 to get the
coefficients, and the coeffs from
http://www.peterstone.name/Maplepgs/Maple/nmthds/RKcoeff/Runge_Kutta_schemes/RK5/RKcoeff5k_3.pdf
. Neither seems to lead to solutions close to the 54 implementations
(hundreds or thousands of km off for a 1 day prop). Something is wrong,
and it's worrisome.

Also attempted different implementations of the adaptive step size, and
so far the WP algorithm is the best: at most four intermediate steps
before hitting the min step size whereas other implementations required
4 or more step changes (some even went into NaN!), while leading to the
same result. We want less iterations for that same result.