Pinpointed the integration error of Fehlberg to the a_coeff algo
Specifically (as seen from the three commented lines of my integration),
it seems that I'm skipping calls to the A matrix. That will definitely
cause an issue.Reading the GMAT source code, what they call the "b"
coefficients are my "a" coeffs (but they seem to add 0.0s at the end of
the row) and their ci and ee are all in my b coeffs as b and b_star. My
hunch is that I'm forgetting to multiply the first dynamics call (i.e.
ki[0]) but the value it should be. I'll read more of the GMAT source
code to check my iteration algorithms.