Demonstration of the combination of dynamics
Two discovered problems:
1. fn state -> &VectorN means that the returned value is borrowed
instead of copied. Although this makes 100% sense for simple dynamics,
it means that when combining dynamics we *must* also have a local copy
of the state of the system. This is annoying because we're effectively
duplicating information in memory at the expense of avoid a copy on
return.

2. In set_state, I cannot seem to pass a split of the states to the
`eom` of each dynamics. This is also annoying because code which could
be literally three lines become 14. I'm sure there's a way, I just
haven't found it yet.