Actually, not fixed, but broken.
I had inadvertantly set the test result to be one from the two body
only. I had a few different results lying around, and I thought I'd
removed that one and kept the previous one. Anyway, that allocation part
of the matrix is important. It also _seems_ to happen only once in GMAT
(but maybe am I reading the code incorrectly). And it seems like it's
reused? That does not make sense to me why that would be the case.