Well this still does not work -- don't know why
I've read the GMAT code for this maybe 20 times now, and rewriting
chunks of it from C++ to Rust, I still can't find where my bug is. The
only last thing I can think of now is that maybe the allocation of the
a_matrix should happen once and we should continue altering that matrix.
But again, this does not make any physical sense to me. I'll probably
check the references given in GMAT (I couldn't find them in 2016a but
found them in the source code of 2018a).