WARNING: This code has a lot of debugging statements
Currently debugging the error handling in the propagator. GMAT R2018a
takes exactly 5 steps to propagator for 60 seconds in a given scenario
with J2, by nyx requires 600 steps because it sets the error threshold
down to 0.1. Maybe I need to increase the step size for the next step,
but weirdly the first step in GMAT is about 28 seconds.