Fixed error computation, but still have a wrong step size with Dormand45
The error was missing a factor. Now, I wonder if the "accuracy" in the
propagator in GMAT is actually different from the tolerance used. With
the same 1e-12 setting, GMAT consistently uses a step size of about 8
seconds, whereas my code which is effectively a "translation" goes down
to 0.1 seconds with a max error on the candidate state on the order
1e-2.