ACTUALLY: the previous code was _mostly_ correct
b_i is the higher order, and b_i_star the local approxiation of lower
order. So I fixed back the code to be what it was before mostly. I still
have a bug when enabling the adaptive step size (error is VERY large
even for two body prop). I checked a million times the step computation
and I cannot find any mistake compared to GMAT (also tried using a i32
instead of a f64 with powf and that broke things by dividing by zero).
So I must have a mistake in the way I compute the relative error.