Added fixed step validation scenarios. Adaptive step size still "broken"
I put broken between quotes because my latest hunch is that the
propagation with an adaptive step size finishes too late! And that would
make sense. For example, with a Verner56, the final adaptive step size
is about 27.59. Hence, we may have gone up to 27.59 seconds past the
deadline time. At 7 km/s, that's a huge difference in state!