Implemented the graph of EXBs. Not yet doing the conversion, soon though
Must rename state because the computation only works for de* files 
(based on the algorithm). This must be made clear from the function 
call. I also need a way to make sure that it isn't called on incorrect 
data -- no clue at all on how to do that for now.