One minus fraction slightly decreases SRP error
I'm a bit surprised: last time I explicitly tested SRP, the error was in the worst case 286 m. Implementing the Montelbruck model decreased that to 121 m, but I hacked that implementation (some NaN in a sine).

Retesting the prior (simpler) implementation has much smaller error, all attributable to the accumulation of error in celestial computation.

Updated the docs.

Closes #206

Signed-off-by: Christopher Rabotin <christopher.rabotin@gmail.com>