Boolean combination of events is unsound w/ Brent
A Brent solve is only for unimodal functions. I implemented a
combination of events method that tried to hack it for several events.
The usage syntax was confusing and complicated, but more importantly, it
did not lead to good results. The `And` combination would work as an
`Or`, and the `Or` combination would just not work. This makes sense in
hindsight because the minima of the product or sum of two unimodals
functions is not necessarily the product or sum of the minima.

That said, this work found and fixed a bug in the finite burn
maneuvering where the maneuver wasn't actually switched off. I'm not
totally sure why the tests didn't catch that, but it's fixed now.

I've also added an example of how to use event findings to match a real
combination of both events.