Fixed remaining tests.
For robustness test, I had to increase the tolerance by an order of magnitude. I explain this because this test uses the new tracking set up. This one uses the actual event of rising above the horizon to find when to start tracking. It causes the OD propagator and the truth propagator to take a different step size.