Remove the xhat dev tests because they were wrong.
Specifically, they would set the initial estimate to be off, but the propagator was still set to the nominal state, so it would only test that the measurement update of the filter would correctly update the state deviation. As such, it did not provide any additional testing compared to the validation tests.